@@ -1,75 +1,31 @@
 {
-  "name": "@oocw/frontend",
-  "version": "0.1.0",
+  "name": "oocw-frontend",
   "private": true,
+  "version": "0.1.0",
+  "type": "module",
   "scripts": {
-    "serve": "vite",
-    "build": "vue-tsc && vite build",
-    "build:client": "vite build --mode=client",
-    "preview": "vite preview --port 5002",
-    "build-lang": "python ../localization/output.py ./src/locales/",
-    "prepare": "cd .. && husky install ./Ra3.battleNet.Frontend/.husky", 
-    "lint": "vue-cli-service lint",
-    "i18n:report": "vue-cli-service i18n:report --src \"./src/**/*.?(js|vue)\" --locales \"./src/locales/**/*.json\""
+    "dev": "vite",
+    "build": "vue-tsc -b && vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@intlify/vue-i18n-loader": "^4.2.0",
-    "animejs": "^3.2.1",
+    "animejs": "^3.2.2",
     "cookie-store": "^3.0.0",
-    "core-js": "^3.26.1",
-    "file-saver": "^2.0.5",
-    "jest": "^29.3.1",
-    "jest-editor-support": "^30.3.1",
-    "lodash": "^4.17.21",
-    "marked": "^4.2.4",
-    "vue": "^3.2.45",
-    "vue-i18n": "^9.2.2",
-    "vue-router": "^4.1.6"
+    "vue": "^3.4.37",
+    "vue-i18n": "^10.0.0",
+    "vue-router": "^4.4.5"
   },
   "devDependencies": {
-    "@intlify/vite-plugin-vue-i18n": "^6.0.3",
-    "@types/animejs": "^3.1.6",
-    "@types/file-saver": "^2.0.5",
-    "@types/lodash": "^4.14.191",
-    "@types/marked": "^4.0.8",
-    "@types/node": "^18.11.12",
-    "@vitejs/plugin-basic-ssl": "^1.0.1",
-    "@vitejs/plugin-vue": "^4.0.0",
-    "eslint": "^8.29.0",
-    "eslint-plugin-vue": "^9.8.0",
-    "husky": "^8.0.2",
-    "lint-staged": "^13.1.0",
-    "prettier": "^2.8.1",
-    "typescript": "^4.9.4",
-    "vite": "^4.0.0",
-    "vite-plugin-pwa": "^0.13.3",
-    "vue-tsc": "^1.0.12"
-  },
-  "lint-staged": {
-    "*.{ts,vue}": "eslint --cache --fix",
-    "*.{ts,vue,css,json,md,html}": "prettier --write"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/vue3-essential",
-      "eslint:recommended",
-      "@vue/typescript"
-    ],
-    "parserOptions": {
-      "parser": "@typescript-eslint/parser"
-    },
-    "rules": {
-      "no-irregular-whitespace": "off"
-    }
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead",
-    "not ie 11"
-  ]
+    "@eslint/js": "^9.11.0",
+    "@types/node": "^22.5.5",
+    "@vitejs/plugin-vue": "^5.1.2",
+    "eslint": "^9.11.0",
+    "eslint-plugin-vue": "^9.28.0",
+    "globals": "^15.9.0",
+    "typescript": "^5.5.3",
+    "typescript-eslint": "^8.6.0",
+    "vite": "^5.4.1",
+    "vite-plugin-mkcert": "^1.17.6",
+    "vue-tsc": "^2.0.29"
+  }
 }
